+++ e2e/commands/add.e2e.js
@@ -180,12 +180,14 @@ describe('bit add command', function () {
     it('Should throw message if adding test files without id', () => {
       helper.createFile('bar', 'foo2.js');
       helper.createFile('bar', 'foo2.spec.js');
-      helper.addComponent(path.normalize('bar/foo2.js'));
-      const addCmd = () => helper.addComponent(` -t ${path.normalize('bar/foo2.spec.js')}`);
+      const normalizedPath = path.normalize('bar/foo2.js');
+      helper.addComponent(normalizedPath);
+      const specNormalizedPath = path.normalize('bar/foo2.spec.js');
+      const addCmd = () => helper.addComponent(` -t ${specNormalizedPath}`);
       expect(addCmd).to.throw(
         `Command failed: ${
           helper.bitBin
-        } add  -t bar/foo2.spec.js\nplease specify a component ID to add test files to an existing component. \nexample: bit add --tests [test_file_path] --id [component_id]\n`
+        } add  -t ${specNormalizedPath}\nplease specify a component ID to add test files to an existing component. \nexample: bit add --tests [test_file_path] --id [component_id]\n`
       );
     });
 
@@ -306,9 +308,7 @@ describe('bit add command', function () {
       expect(addCmd).to.throw(
         `Command failed: ${
           helper.bitBin
-        } add bar -n test\nerror: a main file index.[js, ts, jsx, tsx, css, scss, less, sass] was not found among the component's files ${
-          file1Path
-        }, ${file2Path}. please use 'bit add' --main flag to specify a different main file`
+        } add bar -n test\nerror: a main file index.[js, ts, jsx, tsx, css, scss, less, sass] was not found among the component's files ${file1Path}, ${file2Path}. please use 'bit add' --main flag to specify a different main file`
       );
     });
     it('Should throw error msg if -i and -n flag are used with bit add', () => {
@@ -332,29 +332,32 @@ describe('bit add command', function () {
     it('Should prevent adding a file with invalid keys in namespace', () => {
       let errMsg;
       helper.createComponentBarFoo();
+      const normalizedPath = path.normalize('bar/foo.js');
       try {
-        helper.addComponentWithOptions(path.normalize('bar/foo.js'), { i: 'bar.f/foo' });
+        helper.addComponentWithOptions(normalizedPath, { i: 'bar.f/foo' });
       } catch (err) {
         errMsg = err.message;
       }
       expect(errMsg).to.have.string(
         `Command failed: ${
           helper.bitBin
-        } add bar/foo.js -i bar.f/foo\nerror: "bar.f/foo" is invalid, component IDs can only contain alphanumeric, lowercase characters, and the following ["-", "_", "$", "!"]\n`
+        } add ${normalizedPath} -i bar.f/foo\nerror: "bar.f/foo" is invalid, component IDs can only contain alphanumeric, lowercase characters, and the following ["-", "_", "$", "!"]\n`
       );
     });
     it('Should prevent adding a file with invalid keys in ID', () => {
       let errMsg;
       helper.createComponentBarFoo();
+      let normalizedPath;
       try {
-        helper.addComponentWithOptions(path.normalize('bar/foo.js'), { i: 'bar/fo.o' });
+        normalizedPath = path.normalize('bar/foo.js');
+        helper.addComponentWithOptions(normalizedPath, { i: 'bar/fo.o' });
       } catch (err) {
         errMsg = err.message;
       }
       expect(errMsg).to.have.string(
         `Command failed: ${
           helper.bitBin
-        } add bar/foo.js -i bar/fo.o\nerror: "bar/fo.o" is invalid, component IDs can only contain alphanumeric, lowercase characters, and the following ["-", "_", "$", "!"]\n`
+        } add ${normalizedPath} -i bar/fo.o\nerror: "bar/fo.o" is invalid, component IDs can only contain alphanumeric, lowercase characters, and the following ["-", "_", "$", "!"]\n`
       );
     });
     it.skip('Bitmap mainFile should point to correct mainFile', () => {});
@@ -686,18 +689,21 @@ describe('bit add command', function () {
     beforeEach(() => {
       helper.reInitLocalScope();
     });
-    it('bitMap should not contain component if all files are excluded', () => {
+    it('should throw error when all files are excluded', () => {
       helper.createFile('bar', 'foo1.js');
-      helper.addComponentWithOptions(path.normalize('bar/foo1.js'), { e: 'bar/foo1.js' });
-      const bitMap = helper.readBitMap();
-      expect(bitMap).not.to.have.property('bar/foo1');
+      const normalizedPath = path.normalize('bar/foo1.js');
+      const addCmd = () => helper.addComponentWithOptions(normalizedPath, { e: 'bar/foo1.js' });
+      expect(addCmd).to.throw(`warning: no files to add, the following files were ignored: ${normalizedPath}`);
     });
-    it('bitMap should not contain component if the main file is excluded', () => {
+    it('should throw error when main file is excluded', () => {
       helper.createFile('bar', 'foo1.js');
       helper.createFile('bar', 'foo2.js');
-      helper.addComponentWithOptions('bar', { i: 'bar/foo', e: 'bar/foo1.js', m: 'bar/foo1.js' });
-      const bitMap = helper.readBitMap();
-      expect(bitMap).not.to.have.property('bar/foo');
+      const addCmd = () => helper.addComponentWithOptions('bar', { i: 'bar/foo', e: 'bar/foo1.js', m: 'bar/foo1.js' });
+      expect(addCmd).to.throw(
+        `error: a main file bar/foo1.js was not found among the component's files ${path.normalize(
+          'bar/foo2.js'
+        )}. please use 'bit add' --main flag to specify a different main file`
+      );
     });
     it('bitMap should only contain bits that have files', () => {
       helper.createFile('bar', 'foo1.js');
@@ -731,7 +737,7 @@ describe('bit add command', function () {
       helper.createFile('bar', 'foo1.js');
       helper.createFile('bar', 'foo2.js');
       helper.createFile('bar/x', 'foo2.exclude.js');
-      helper.addComponentWithOptions('bar/*', { e: 'bar/x/*' });
+      helper.addComponentWithOptions('bar/*', { e: 'bar/x/' });
       const bitMap = helper.readBitMap();
       expect(bitMap).to.have.property('bar/foo1');
       expect(bitMap).to.have.property('bar/foo2');
@@ -766,6 +772,16 @@ describe('bit add command', function () {
       const files = bitMap['bar/foo1'].files;
       expect(bitMap).to.have.property('bar/foo1');
       expect(files).to.be.ofSize(1);
+    });
+    it('bit should add components and exclude files', () => {
+      helper.createFile('bar', 'foo1.js');
+      helper.createFile('bar', 'foo1.js');
+      helper.createFile('bar', 'index.js');
+      helper.createFile('foo', 'foo3.js');
+      helper.createFile('foo', 'foo4.js');
+      helper.addComponentWithOptions(path.normalize('*'), { e: 'foo' });
+      const bitMap = helper.readBitMap();
+      expect(bitMap).not.to.have.property('bar/foo1');
     });
   });
   describe('with multiple index files', () => {
@@ -809,9 +825,9 @@ describe('bit add command', function () {
       it('should throw an error', () => {
         const barFoo2Path = path.join('bar', 'foo2.js');
         expect(output).to.have.string(
-          `Command failed: ${helper.bitBin} add ${
-            barFoo2Path
-          } -i bar/foo\nunable to add file bar/foo2.js because it's located outside the component root dir components/bar/foo\n`
+          `Command failed: ${
+            helper.bitBin
+          } add ${barFoo2Path} -i bar/foo\nunable to add file bar/foo2.js because it's located outside the component root dir components/bar/foo\n`
         );
       });
     });